refactor(ProductDao): 의미없는 인터페이스 제거
