+++ src/main/java/cart/dao/ProductDao.java
@@ -1,6 +1,7 @@
 package cart.dao;
 
-import cart.controller.ProductRequest;
+import cart.controller.dto.ProductRequest;
+import cart.dao.entity.ProductEntity;
 import java.sql.PreparedStatement;
 import java.util.List;
 import java.util.Optional;
@@ -10,7 +11,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class ProductDao implements CrudDao<ProductEntity, ProductRequest> {
+public class ProductDao {
 
     private final JdbcTemplate jdbcTemplate;
 
@@ -18,7 +19,6 @@ public class ProductDao implements CrudDao<ProductEntity, ProductRequest> {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    @Override
     public long add(ProductRequest request) {
         String query = "INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)";
         GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
@@ -32,7 +32,6 @@ public class ProductDao implements CrudDao<ProductEntity, ProductRequest> {
         return keyHolder.getKey().longValue();
     }
 
-    @Override
     public List<ProductEntity> findAll() {
         String query = "SELECT * FROM product";
         return jdbcTemplate.query(query, (resultSet, rowNum) ->
@@ -43,23 +42,20 @@ public class ProductDao implements CrudDao<ProductEntity, ProductRequest> {
                 resultSet.getString("image_url")));
     }
 
-    @Override
     public Optional<ProductEntity> findById(Long id) {
         String query = "SELECT * FROM product WHERE id = ?";
-        try{
+        try {
             return Optional.ofNullable(jdbcTemplate.queryForObject(query, (resultSet, rowNum) ->
                 new ProductEntity(
                     resultSet.getLong("id"),
                     resultSet.getString("name"),
                     resultSet.getInt("price"),
                     resultSet.getString("image_url")), id));
-        } catch (EmptyResultDataAccessException exception){
+        } catch (EmptyResultDataAccessException exception) {
             return Optional.empty();
         }
-
     }
 
-    @Override
     public int updateById(Long id, ProductRequest request) {
         String query = "UPDATE product SET name = ?, price = ?, image_url = ? WHERE id = ?";
         return jdbcTemplate.update(query, request.getName(), request.getPrice(),
@@ -67,7 +63,6 @@ public class ProductDao implements CrudDao<ProductEntity, ProductRequest> {
             id);
     }
 
-    @Override
     public int deleteById(Long id) {
         String query = "DELETE FROM product WHERE id = ?";
         return jdbcTemplate.update(query, id);
